Fix errors with creating project with owner
Create `objectInArray` to return index of object that matches the
key/value pair. Replaced `projectsInArray` function with this as
well to clean up some code (hopefully). `project create` also
works by finding any org email that the user is part of instead of
only working if email is the last org in the list.

[MODULUS-1590]